requirements.txt
langchain-google-genai
langchain
langgraph
python-dotenv
tavily-python
pandas
faiss-cpu
sentence-transformers
langchain-community
unstructured
langchain-huggingface